load plugin only in linux
