+++ plugin_other.go
@@ -1,0 +1,7 @@
+// +build !linux
+
+package core
+
+func LoadPlugins() error {
+	return nil
+}
